Refactor adapter.rs into multiple files.
