+++ src/adapter/mod.rs
@@ -1,20 +1,25 @@
-use std::{collections::BTreeSet, rc::Rc, sync::Arc};
+use std::{collections::BTreeSet, sync::Arc};
 
-use rustdoc_types::{
-    Crate, Enum, Function, Id, Impl, Item, ItemEnum, Path, Span, Struct, Trait, Type, Variant,
-    VariantKind,
-};
+use rustdoc_types::{Id, ItemEnum, VariantKind};
 use trustfall::{
     provider::{
-        accessor_property, field_property, resolve_coercion_with, resolve_neighbors_with,
-        resolve_property_with, Adapter, ContextIterator, ContextOutcomeIterator, EdgeParameters,
-        ResolveEdgeInfo, ResolveInfo, Typename, VertexIterator,
+        resolve_coercion_with, resolve_neighbors_with, Adapter, ContextIterator,
+        ContextOutcomeIterator, EdgeParameters, ResolveEdgeInfo, ResolveInfo, Typename,
+        VertexIterator,
     },
     FieldValue, Schema,
 };
 
-use crate::attributes::{Attribute, AttributeMetaItem};
-use crate::indexed_crate::IndexedCrate;
+use crate::{attributes::Attribute, indexed_crate::IndexedCrate};
+
+use self::{
+    origin::Origin,
+    vertex::{Vertex, VertexKind},
+};
+
+mod origin;
+mod properties;
+mod vertex;
 
 #[non_exhaustive]
 pub struct RustdocAdapter<'a> {
@@ -34,576 +39,7 @@ impl<'a> RustdocAdapter<'a> {
     }
 
     pub fn schema() -> Schema {
-        Schema::parse(include_str!("rustdoc_schema.graphql")).expect("schema not valid")
-    }
-}
-
-#[non_exhaustive]
-#[derive(Debug, Clone, Copy)]
-pub enum Origin {
-    CurrentCrate,
-    PreviousCrate,
-}
-
-impl Origin {
-    fn make_item_vertex<'a>(&self, item: &'a Item) -> Vertex<'a> {
-        Vertex {
-            origin: *self,
-            kind: item.into(),
-        }
-    }
-
-    fn make_span_vertex<'a>(&self, span: &'a Span) -> Vertex<'a> {
-        Vertex {
-            origin: *self,
-            kind: span.into(),
-        }
-    }
-
-    fn make_path_vertex<'a>(&self, path: &'a [String]) -> Vertex<'a> {
-        Vertex {
-            origin: *self,
-            kind: VertexKind::Path(path),
-        }
-    }
-
-    fn make_importable_path_vertex<'a>(&self, importable_path: Vec<&'a str>) -> Vertex<'a> {
-        Vertex {
-            origin: *self,
-            kind: VertexKind::ImportablePath(importable_path),
-        }
-    }
-
-    fn make_raw_type_vertex<'a>(&self, raw_type: &'a rustdoc_types::Type) -> Vertex<'a> {
-        Vertex {
-            origin: *self,
-            kind: VertexKind::RawType(raw_type),
-        }
-    }
-
-    fn make_attribute_vertex<'a>(&self, attr: Attribute<'a>) -> Vertex<'a> {
-        Vertex {
-            origin: *self,
-            kind: VertexKind::Attribute(attr),
-        }
-    }
-
-    fn make_attribute_meta_item_vertex<'a>(
-        &self,
-        meta_item: Rc<AttributeMetaItem<'a>>,
-    ) -> Vertex<'a> {
-        Vertex {
-            origin: *self,
-            kind: VertexKind::AttributeMetaItem(meta_item),
-        }
-    }
-
-    fn make_implemented_trait_vertex<'a>(
-        &self,
-        path: &'a rustdoc_types::Path,
-        trait_def: &'a Item,
-    ) -> Vertex<'a> {
-        Vertex {
-            origin: *self,
-            kind: VertexKind::ImplementedTrait(path, trait_def),
-        }
-    }
-
-    fn make_function_parameter_vertex<'a>(&self, name: &'a str) -> Vertex<'a> {
-        Vertex {
-            origin: *self,
-            kind: VertexKind::FunctionParameter(name),
-        }
-    }
-}
-
-#[non_exhaustive]
-#[derive(Debug, Clone)]
-pub struct Vertex<'a> {
-    origin: Origin,
-    kind: VertexKind<'a>,
-}
-
-#[non_exhaustive]
-#[derive(Debug, Clone)]
-pub enum VertexKind<'a> {
-    CrateDiff((&'a IndexedCrate<'a>, &'a IndexedCrate<'a>)),
-    Crate(&'a IndexedCrate<'a>),
-    Item(&'a Item),
-    Span(&'a Span),
-    Path(&'a [String]),
-    ImportablePath(Vec<&'a str>),
-    RawType(&'a Type),
-    Attribute(Attribute<'a>),
-    AttributeMetaItem(Rc<AttributeMetaItem<'a>>),
-    ImplementedTrait(&'a Path, &'a Item),
-    FunctionParameter(&'a str),
-}
-
-impl<'a> Typename for Vertex<'a> {
-    /// The name of the actual runtime type of this vertex,
-    /// intended to fulfill resolution requests for the __typename property.
-    #[inline]
-    fn typename(&self) -> &'static str {
-        match self.kind {
-            VertexKind::Item(item) => match &item.inner {
-                rustdoc_types::ItemEnum::Struct(..) => "Struct",
-                rustdoc_types::ItemEnum::Enum(..) => "Enum",
-                rustdoc_types::ItemEnum::Function(..) => "Function",
-                rustdoc_types::ItemEnum::Variant(variant) => match variant.kind {
-                    VariantKind::Plain => "PlainVariant",
-                    VariantKind::Tuple(..) => "TupleVariant",
-                    VariantKind::Struct { .. } => "StructVariant",
-                },
-                rustdoc_types::ItemEnum::StructField(..) => "StructField",
-                rustdoc_types::ItemEnum::Impl(..) => "Impl",
-                rustdoc_types::ItemEnum::Trait(..) => "Trait",
-                _ => unreachable!("unexpected item.inner for item: {item:?}"),
-            },
-            VertexKind::Span(..) => "Span",
-            VertexKind::Path(..) => "Path",
-            VertexKind::ImportablePath(..) => "ImportablePath",
-            VertexKind::Crate(..) => "Crate",
-            VertexKind::CrateDiff(..) => "CrateDiff",
-            VertexKind::Attribute(..) => "Attribute",
-            VertexKind::AttributeMetaItem(..) => "AttributeMetaItem",
-            VertexKind::ImplementedTrait(..) => "ImplementedTrait",
-            VertexKind::RawType(ty) => match ty {
-                rustdoc_types::Type::ResolvedPath { .. } => "ResolvedPathType",
-                rustdoc_types::Type::Primitive(..) => "PrimitiveType",
-                _ => "OtherType",
-            },
-            VertexKind::FunctionParameter(..) => "FunctionParameter",
-        }
-    }
-}
-
-#[allow(dead_code)]
-impl<'a> Vertex<'a> {
-    fn new_crate(origin: Origin, crate_: &'a IndexedCrate<'a>) -> Self {
-        Self {
-            origin,
-            kind: VertexKind::Crate(crate_),
-        }
-    }
-
-    fn as_crate_diff(&self) -> Option<(&'a IndexedCrate<'a>, &'a IndexedCrate<'a>)> {
-        match &self.kind {
-            VertexKind::CrateDiff(tuple) => Some(*tuple),
-            _ => None,
-        }
-    }
-
-    fn as_indexed_crate(&self) -> Option<&'a IndexedCrate<'a>> {
-        match self.kind {
-            VertexKind::Crate(c) => Some(c),
-            _ => None,
-        }
-    }
-
-    fn as_crate(&self) -> Option<&'a Crate> {
-        self.as_indexed_crate().map(|c| c.inner)
-    }
-
-    fn as_item(&self) -> Option<&'a Item> {
-        match self.kind {
-            VertexKind::Item(item) => Some(item),
-            _ => None,
-        }
-    }
-
-    fn as_struct(&self) -> Option<&'a Struct> {
-        self.as_item().and_then(|item| match &item.inner {
-            rustdoc_types::ItemEnum::Struct(s) => Some(s),
-            _ => None,
-        })
-    }
-
-    fn as_struct_field(&self) -> Option<&'a Type> {
-        self.as_item().and_then(|item| match &item.inner {
-            rustdoc_types::ItemEnum::StructField(s) => Some(s),
-            _ => None,
-        })
-    }
-
-    fn as_span(&self) -> Option<&'a Span> {
-        match self.kind {
-            VertexKind::Span(s) => Some(s),
-            _ => None,
-        }
-    }
-
-    fn as_enum(&self) -> Option<&'a Enum> {
-        self.as_item().and_then(|item| match &item.inner {
-            rustdoc_types::ItemEnum::Enum(e) => Some(e),
-            _ => None,
-        })
-    }
-
-    fn as_trait(&self) -> Option<&'a Trait> {
-        self.as_item().and_then(|item| match &item.inner {
-            rustdoc_types::ItemEnum::Trait(t) => Some(t),
-            _ => None,
-        })
-    }
-
-    fn as_variant(&self) -> Option<&'a Variant> {
-        self.as_item().and_then(|item| match &item.inner {
-            rustdoc_types::ItemEnum::Variant(v) => Some(v),
-            _ => None,
-        })
-    }
-
-    fn as_path(&self) -> Option<&'a [String]> {
-        match &self.kind {
-            VertexKind::Path(path) => Some(*path),
-            _ => None,
-        }
-    }
-
-    fn as_importable_path(&self) -> Option<&'_ Vec<&'a str>> {
-        match &self.kind {
-            VertexKind::ImportablePath(path) => Some(path),
-            _ => None,
-        }
-    }
-
-    fn as_function(&self) -> Option<&'a Function> {
-        self.as_item().and_then(|item| match &item.inner {
-            rustdoc_types::ItemEnum::Function(func) => Some(func),
-            _ => None,
-        })
-    }
-
-    fn as_function_parameter(&self) -> Option<&'a str> {
-        match &self.kind {
-            VertexKind::FunctionParameter(name) => Some(name),
-            _ => None,
-        }
-    }
-
-    fn as_impl(&self) -> Option<&'a Impl> {
-        self.as_item().and_then(|item| match &item.inner {
-            rustdoc_types::ItemEnum::Impl(x) => Some(x),
-            _ => None,
-        })
-    }
-
-    fn as_attribute(&self) -> Option<&'_ Attribute<'a>> {
-        match &self.kind {
-            VertexKind::Attribute(attr) => Some(attr),
-            _ => None,
-        }
-    }
-
-    fn as_attribute_meta_item(&self) -> Option<&'_ AttributeMetaItem<'a>> {
-        match &self.kind {
-            VertexKind::AttributeMetaItem(meta_item) => Some(meta_item),
-            _ => None,
-        }
-    }
-
-    fn as_raw_type(&self) -> Option<&'a rustdoc_types::Type> {
-        match &self.kind {
-            VertexKind::RawType(ty) => Some(*ty),
-            _ => None,
-        }
-    }
-
-    fn as_implemented_trait(&self) -> Option<(&'a rustdoc_types::Path, &'a Item)> {
-        match &self.kind {
-            VertexKind::ImplementedTrait(path, trait_item) => Some((*path, *trait_item)),
-            _ => None,
-        }
-    }
-}
-
-impl<'a> From<&'a Item> for VertexKind<'a> {
-    fn from(item: &'a Item) -> Self {
-        Self::Item(item)
-    }
-}
-
-impl<'a> From<&'a IndexedCrate<'a>> for VertexKind<'a> {
-    fn from(c: &'a IndexedCrate<'a>) -> Self {
-        Self::Crate(c)
-    }
-}
-
-impl<'a> From<&'a Span> for VertexKind<'a> {
-    fn from(s: &'a Span) -> Self {
-        Self::Span(s)
-    }
-}
-
-fn resolve_crate_property<'a>(
-    contexts: ContextIterator<'a, Vertex<'a>>,
-    property_name: &str,
-) -> ContextOutcomeIterator<'a, Vertex<'a>, FieldValue> {
-    match property_name {
-        "root" => resolve_property_with(
-            contexts,
-            field_property!(as_crate, root, { root.0.clone().into() }),
-        ),
-        "crate_version" => {
-            resolve_property_with(contexts, field_property!(as_crate, crate_version))
-        }
-        "includes_private" => {
-            resolve_property_with(contexts, field_property!(as_crate, includes_private))
-        }
-        "format_version" => {
-            resolve_property_with(contexts, field_property!(as_crate, format_version))
-        }
-        _ => unreachable!("Crate property {property_name}"),
-    }
-}
-
-fn resolve_item_property<'a>(
-    contexts: ContextIterator<'a, Vertex<'a>>,
-    property_name: &str,
-) -> ContextOutcomeIterator<'a, Vertex<'a>, FieldValue> {
-    match property_name {
-        "id" => resolve_property_with(
-            contexts,
-            field_property!(as_item, id, { id.0.clone().into() }),
-        ),
-        "crate_id" => resolve_property_with(contexts, field_property!(as_item, crate_id)),
-        "name" => resolve_property_with(contexts, field_property!(as_item, name)),
-        "docs" => resolve_property_with(contexts, field_property!(as_item, docs)),
-        "attrs" => resolve_property_with(contexts, field_property!(as_item, attrs)),
-        "visibility_limit" => resolve_property_with(contexts, |vertex| {
-            let item = vertex.as_item().expect("not an item");
-            match &item.visibility {
-                rustdoc_types::Visibility::Public => "public".into(),
-                rustdoc_types::Visibility::Default => "default".into(),
-                rustdoc_types::Visibility::Crate => "crate".into(),
-                rustdoc_types::Visibility::Restricted { parent: _, path } => {
-                    format!("restricted ({path})").into()
-                }
-            }
-        }),
-        _ => unreachable!("Item property {property_name}"),
-    }
-}
-
-fn resolve_struct_property<'a>(
-    contexts: ContextIterator<'a, Vertex<'a>>,
-    property_name: &str,
-) -> ContextOutcomeIterator<'a, Vertex<'a>, FieldValue> {
-    match property_name {
-        "struct_type" => resolve_property_with(contexts, |vertex| {
-            let struct_vertex = vertex.as_struct().expect("not a struct");
-            match struct_vertex.kind {
-                rustdoc_types::StructKind::Plain { .. } => "plain",
-                rustdoc_types::StructKind::Tuple(..) => "tuple",
-                rustdoc_types::StructKind::Unit => "unit",
-            }
-            .into()
-        }),
-        "fields_stripped" => resolve_property_with(contexts, |vertex| {
-            let struct_vertex = vertex.as_struct().expect("not a struct");
-            match struct_vertex.kind {
-                rustdoc_types::StructKind::Plain {
-                    fields_stripped, ..
-                } => fields_stripped.into(),
-                _ => FieldValue::Null,
-            }
-        }),
-        _ => unreachable!("Struct property {property_name}"),
-    }
-}
-
-fn resolve_span_property<'a>(
-    contexts: ContextIterator<'a, Vertex<'a>>,
-    property_name: &str,
-) -> ContextOutcomeIterator<'a, Vertex<'a>, FieldValue> {
-    match property_name {
-        "filename" => resolve_property_with(
-            contexts,
-            field_property!(as_span, filename, {
-                filename.to_str().expect("non-representable path").into()
-            }),
-        ),
-        "begin_line" => resolve_property_with(
-            contexts,
-            field_property!(as_span, begin, { (begin.0 as u64).into() }),
-        ),
-        "begin_column" => resolve_property_with(
-            contexts,
-            field_property!(as_span, begin, { (begin.1 as u64).into() }),
-        ),
-        "end_line" => resolve_property_with(
-            contexts,
-            field_property!(as_span, end, { (end.0 as u64).into() }),
-        ),
-        "end_column" => resolve_property_with(
-            contexts,
-            field_property!(as_span, end, { (end.1 as u64).into() }),
-        ),
-        _ => unreachable!("Span property {property_name}"),
-    }
-}
-
-fn resolve_enum_property<'a>(
-    contexts: ContextIterator<'a, Vertex<'a>>,
-    property_name: &str,
-) -> ContextOutcomeIterator<'a, Vertex<'a>, FieldValue> {
-    match property_name {
-        "variants_stripped" => {
-            resolve_property_with(contexts, field_property!(as_enum, variants_stripped))
-        }
-        _ => unreachable!("Enum property {property_name}"),
-    }
-}
-
-fn resolve_path_property<'a>(
-    contexts: ContextIterator<'a, Vertex<'a>>,
-    property_name: &str,
-) -> ContextOutcomeIterator<'a, Vertex<'a>, FieldValue> {
-    match property_name {
-        "path" => resolve_property_with(contexts, |vertex| {
-            vertex.as_path().expect("not a path").into()
-        }),
-        _ => unreachable!("Path property {property_name}"),
-    }
-}
-
-fn resolve_importable_path_property<'a>(
-    contexts: ContextIterator<'a, Vertex<'a>>,
-    property_name: &str,
-) -> ContextOutcomeIterator<'a, Vertex<'a>, FieldValue> {
-    match property_name {
-        "path" => resolve_property_with(contexts, |vertex| {
-            vertex
-                .as_importable_path()
-                .expect("not an importable path")
-                .iter()
-                .map(ToString::to_string)
-                .collect::<Vec<_>>()
-                .into()
-        }),
-        "visibility_limit" => resolve_property_with(contexts, |_| "public".into()),
-        _ => unreachable!("ImportablePath property {property_name}"),
-    }
-}
-
-fn resolve_function_like_property<'a>(
-    contexts: ContextIterator<'a, Vertex<'a>>,
-    property_name: &str,
-) -> ContextOutcomeIterator<'a, Vertex<'a>, FieldValue> {
-    match property_name {
-        "const" => resolve_property_with(
-            contexts,
-            field_property!(as_function, header, { header.const_.into() }),
-        ),
-        "async" => resolve_property_with(
-            contexts,
-            field_property!(as_function, header, { header.async_.into() }),
-        ),
-        "unsafe" => resolve_property_with(
-            contexts,
-            field_property!(as_function, header, { header.unsafe_.into() }),
-        ),
-        _ => unreachable!("FunctionLike property {property_name}"),
-    }
-}
-
-fn resolve_function_parameter_property<'a>(
-    contexts: ContextIterator<'a, Vertex<'a>>,
-    property_name: &str,
-) -> ContextOutcomeIterator<'a, Vertex<'a>, FieldValue> {
-    match property_name {
-        "name" => resolve_property_with(contexts, |vertex| {
-            vertex
-                .as_function_parameter()
-                .expect("not a function parameter")
-                .into()
-        }),
-        _ => unreachable!("FunctionParameter property {property_name}"),
-    }
-}
-
-fn resolve_impl_property<'a>(
-    contexts: ContextIterator<'a, Vertex<'a>>,
-    property_name: &str,
-) -> ContextOutcomeIterator<'a, Vertex<'a>, FieldValue> {
-    match property_name {
-        "unsafe" => resolve_property_with(contexts, field_property!(as_impl, is_unsafe)),
-        "negative" => resolve_property_with(contexts, field_property!(as_impl, negative)),
-        "synthetic" => resolve_property_with(contexts, field_property!(as_impl, synthetic)),
-        _ => unreachable!("Impl property {property_name}"),
-    }
-}
-
-fn resolve_attribute_property<'a>(
-    contexts: ContextIterator<'a, Vertex<'a>>,
-    property_name: &str,
-) -> ContextOutcomeIterator<'a, Vertex<'a>, FieldValue> {
-    match property_name {
-        "raw_attribute" => {
-            resolve_property_with(contexts, accessor_property!(as_attribute, raw_attribute))
-        }
-        "is_inner" => resolve_property_with(contexts, field_property!(as_attribute, is_inner)),
-        _ => unreachable!("Attribute property {property_name}"),
-    }
-}
-
-fn resolve_attribute_meta_item_property<'a>(
-    contexts: ContextIterator<'a, Vertex<'a>>,
-    property_name: &str,
-) -> ContextOutcomeIterator<'a, Vertex<'a>, FieldValue> {
-    match property_name {
-        "raw_item" => {
-            resolve_property_with(contexts, field_property!(as_attribute_meta_item, raw_item))
-        }
-        "base" => resolve_property_with(contexts, field_property!(as_attribute_meta_item, base)),
-        "assigned_item" => resolve_property_with(
-            contexts,
-            field_property!(as_attribute_meta_item, assigned_item),
-        ),
-        _ => unreachable!("AttributeMetaItem property {property_name}"),
-    }
-}
-
-fn resolve_raw_type_property<'a>(
-    contexts: ContextIterator<'a, Vertex<'a>>,
-    property_name: &str,
-) -> ContextOutcomeIterator<'a, Vertex<'a>, FieldValue> {
-    match property_name {
-        "name" => resolve_property_with(contexts, |vertex| {
-            let type_vertex = vertex.as_raw_type().expect("not a RawType");
-            match type_vertex {
-                rustdoc_types::Type::ResolvedPath(path) => path.name.clone().into(),
-                rustdoc_types::Type::Primitive(name) => name.clone().into(),
-                _ => unreachable!("unexpected RawType vertex content: {type_vertex:?}"),
-            }
-        }),
-        _ => unreachable!("RawType property {property_name}"),
-    }
-}
-
-fn resolve_trait_property<'a>(
-    contexts: ContextIterator<'a, Vertex<'a>>,
-    property_name: &str,
-) -> ContextOutcomeIterator<'a, Vertex<'a>, FieldValue> {
-    match property_name {
-        "unsafe" => resolve_property_with(contexts, field_property!(as_trait, is_unsafe)),
-        _ => unreachable!("Trait property {property_name}"),
-    }
-}
-
-fn resolve_implemented_trait_property<'a>(
-    contexts: ContextIterator<'a, Vertex<'a>>,
-    property_name: &str,
-) -> ContextOutcomeIterator<'a, Vertex<'a>, FieldValue> {
-    match property_name {
-        "name" => resolve_property_with(contexts, |vertex| {
-            let (path, _) = vertex
-                .as_implemented_trait()
-                .expect("not an ImplementedTrait");
-            path.name.clone().into()
-        }),
-        _ => unreachable!("Trait property {property_name}"),
+        Schema::parse(include_str!("../rustdoc_schema.graphql")).expect("schema not valid")
     }
 }
 
@@ -649,8 +85,8 @@ impl<'a> Adapter<'a> for RustdocAdapter<'a> {
             }))
         } else {
             match type_name.as_ref() {
-                "Crate" => resolve_crate_property(contexts, property_name),
-                "Item" => resolve_item_property(contexts, property_name),
+                "Crate" => properties::resolve_crate_property(contexts, property_name),
+                "Item" => properties::resolve_item_property(contexts, property_name),
                 "ImplOwner" | "Struct" | "StructField" | "Enum" | "Variant" | "PlainVariant"
                 | "TupleVariant" | "StructVariant" | "Trait" | "Function" | "Method" | "Impl"
                     if matches!(
@@ -659,31 +95,37 @@ impl<'a> Adapter<'a> for RustdocAdapter<'a> {
                     ) =>
                 {
                     // properties inherited from Item, accesssed on Item subtypes
-                    resolve_item_property(contexts, property_name)
+                    properties::resolve_item_property(contexts, property_name)
                 }
-                "Struct" => resolve_struct_property(contexts, property_name),
-                "Enum" => resolve_enum_property(contexts, property_name),
-                "Span" => resolve_span_property(contexts, property_name),
-                "Path" => resolve_path_property(contexts, property_name),
-                "ImportablePath" => resolve_importable_path_property(contexts, property_name),
+                "Struct" => properties::resolve_struct_property(contexts, property_name),
+                "Enum" => properties::resolve_enum_property(contexts, property_name),
+                "Span" => properties::resolve_span_property(contexts, property_name),
+                "Path" => properties::resolve_path_property(contexts, property_name),
+                "ImportablePath" => {
+                    properties::resolve_importable_path_property(contexts, property_name)
+                }
                 "FunctionLike" | "Function" | "Method"
                     if matches!(property_name.as_ref(), "const" | "unsafe" | "async") =>
                 {
-                    resolve_function_like_property(contexts, property_name)
+                    properties::resolve_function_like_property(contexts, property_name)
                 }
-                "FunctionParameter" => resolve_function_parameter_property(contexts, property_name),
-                "Impl" => resolve_impl_property(contexts, property_name),
-                "Attribute" => resolve_attribute_property(contexts, property_name),
+                "FunctionParameter" => {
+                    properties::resolve_function_parameter_property(contexts, property_name)
+                }
+                "Impl" => properties::resolve_impl_property(contexts, property_name),
+                "Attribute" => properties::resolve_attribute_property(contexts, property_name),
                 "AttributeMetaItem" => {
-                    resolve_attribute_meta_item_property(contexts, property_name)
+                    properties::resolve_attribute_meta_item_property(contexts, property_name)
                 }
-                "Trait" => resolve_trait_property(contexts, property_name),
-                "ImplementedTrait" => resolve_implemented_trait_property(contexts, property_name),
+                "Trait" => properties::resolve_trait_property(contexts, property_name),
+                "ImplementedTrait" => {
+                    properties::resolve_implemented_trait_property(contexts, property_name)
+                }
                 "RawType" | "ResolvedPathType" | "PrimitiveType"
                     if matches!(property_name.as_ref(), "name") =>
                 {
                     // fields from "RawType"
-                    resolve_raw_type_property(contexts, property_name)
+                    properties::resolve_raw_type_property(contexts, property_name)
                 }
                 _ => unreachable!("resolve_property {type_name} {property_name}"),
             }
